Did some refactoring as well as as visualized false negatives
Looking at the precision v recall plot, it seems that recall is quite
low. This means that we have a high % of false negatives. Visualizing the
ground truth, it seems that a lot of them are actually wrong (mislabelled
as rest pose).

This may be because using 10 poses to make a decision may be a lot.